refactor: 노선 읽기 전용 DTO 이름 변경
