+++ src/main/resources/schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS line
 (
     id BIGINT AUTO_INCREMENT NOT NULL,
     name VARCHAR(255) NOT NULL UNIQUE,
-    color VARCHAR(255) NOT NULL UNIQUE,
+    color VARCHAR(255) NOT NULL,
     PRIMARY KEY(id)
 );
